+++ Journal-Limpet/Pages/Journals.cshtml.cs
@@ -27,17 +27,17 @@ namespace Journal_Limpet.Pages
 
         public class JournalListItem
         {
-            public int JournalId { get; set; }
+            public long JournalId { get; set; }
             public string JournalDate { get; set; }
             public string CompleteEntry { get; set; }
-            public int NumberOfLines { get; set; }
+            public long NumberOfLines { get; set; }
 
             public JournalListItem(DataRow row)
             {
-                JournalId = row.Field<int>("journal_id");
+                JournalId = row.Field<long>("journal_id");
                 JournalDate = new DateTimeOffset(row.Field<DateTime>("journal_date"), TimeSpan.Zero).Date.ToString("yyyy-MM-dd");
                 CompleteEntry = row.Field<bool>("complete_entry") ? "Yes" : "No";
-                NumberOfLines = row.Field<int>("last_processed_line_number");
+                NumberOfLines = row.Field<long>("last_processed_line_number");
             }
         }
     }
